Move errors and associated messages to new folder to allow importing from anywhere in pyccel
