+++ pyccel/codegen/codegen.py
@@ -5,11 +5,11 @@ from pyccel.codegen.printing import fcode, ccode, pycode
 
 from pyccel.ast import FunctionDef, Module, Program, Interface
 from pyccel.ast import Header, EmptyLine, NewLine, Comment, CommentBlock
-from pyccel.parser.errors import Errors
+from pyccel.errors.errors import Errors
 
 # TODO improve this import
 
-from pyccel.parser.messages import *
+from pyccel.errors.messages import *
 
 _extension_registry = {'fortran': 'f90', 'c':'c',  'python':'py'}
 printer_registry    = {'fortran':fcode, 'c':ccode, 'python':pycode}
