+++ pyccel/codegen/printing/codeprinter.py
@@ -14,7 +14,7 @@ from sympy.printing.str import StrPrinter
 from pyccel.ast.core import Assign
 from pyccel.ast import Real
 
-from pyccel.parser.errors     import Errors
+from pyccel.errors.errors     import Errors
 
 # TODO: add examples
