+++ pyccel/errors/__init__.py
@@ -1,0 +1,2 @@
+from .errors import *
+from .messages import *
