+++ pyccel/parser/__init__.py
@@ -5,5 +5,3 @@ from .syntax    import *
 from .syntactic import *
 from .semantic  import *
 from .parser    import *
-from .errors    import *
-from .messages  import *
