+++ tests/symbolic/test_symbolic.py
@@ -2,7 +2,7 @@
 
 from pyccel.parser import Parser
 from pyccel.codegen import Codegen
-from pyccel.parser.errors import Errors
+from pyccel.errors.errors import Errors
 import os
 import pytest
