+++ tests/syntax/test_syntax.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 
 from pyccel.parser import Parser
-from pyccel.parser.errors import Errors
+from pyccel.errors.errors import Errors
 import os
 import pytest
